feat: refactored NewsTableViewController into several extensions
